+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -18,10 +18,12 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
 
     assert!(data.members_migrated_to_stable_memory);
+
+    data.events.migrate_to_stable_memory();
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);